+++ .gitignore
@@ -1,10 +1,43 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+# Serverless
+.serverless
+.env
+tmp
+.coveralls.yml
 
-# dependencies
-/node_modules
+# Google
+keyfile.json
+
+# Logs
+*.log
+
+# Directory for instrumented libs generated by jscoverage/JSCover
+lib-cov
+
+# Coverage directory used by tools like istanbul
+coverage
+
+# Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-task-files)
+.grunt
+
+# node-waf configuration
+.lock-wscript
+
+# Compiled binary addons (http://nodejs.org/api/addons.html)
+build/Release
+
+# Dependency directory
+# https://www.npmjs.org/doc/misc/npm-faq.html#should-i-check-my-node_modules-folder-into-git
+node_modules
 /.pnp
 .pnp.js
 .vscode
+
+# IDE
+**/.idea
+
+# OS
+.DS_Store
+.tmp
 
 # testing
 /coverage
@@ -13,7 +46,6 @@
 /build
 
 # misc
-.DS_Store
 .env.local
 .env.development.local
 .env.test.local
