+++ main.py
@@ -1,11 +1,11 @@
-import functions
-from flask import jsonify
+def http(request):
+    """Responds to any HTTP request.
+    Args:
+        request (flask.Request): HTTP request object.
+    Returns:
+        The response text or any set of values that can be turned into a
+        Response object using
+        `make_response <http://flask.pocoo.org/docs/1.0/api/#flask.Flask.make_response>`.
+    """
+    return f'Hello World!'
 
-
-def search_data(request):
-    response_body = functions.search_data(request)
-
-    response = jsonify(response_body)
-    response.headers.add('Access-Control-Allow-Origin', '*')
-
-    return response
